mrc-ide-covidsim: Refactor param file serialization into its own file.
This is the first step in trying to make our parameter generation code
more modular and testable. There isn't really a reason at this point for
the serialization code to be mixed in with the code that generates the
parameter values.
