+++ packages/mrc-ide-covidsim/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@covid-modeling/mrc-ide-covidsim",
-  "version": "1.7.3",
+  "version": "1.8.0",
   "main": "dist/index.js",
   "scripts": {
     "test": "PATH=$PATH:../../node_modules/.bin && mocha --debug-brk --ui tdd -r ts-node/register test/unit/*-test.ts",
