+++ packages/mrc-ide-covidsim/test/unit/imperial-params-test.ts
@@ -1,141 +1,9 @@
 import { assert } from 'chai'
 import * as params from '../../src/imperial-params'
-import { parse, serialize } from '../../src/imperial-params'
+import * as paramsSerialization from '../../src/params-serialization'
 import { input } from '@covid-modeling/api'
 
 suite('the imperial model parameter format', () => {
-  test('parse parameter file', () => {
-    const testParams = parse(`
-[a number]
-1.0
-^^ a comment
-
-//// A second comment
-[an array]
-1.0\t2.0\t3.0
-
-=====
-
-[a number in a new section]
-1
-
-*** ^^ another comment
-
-[a matrix]
-1\t2\t3
-4\t5\t6
-`)
-
-    assert.deepEqual(testParams, {
-      'a number': 1.0,
-      'an array': [1.0, 2.0, 3.0],
-      'a number in a new section': 1,
-      'a matrix': [
-        [1, 2, 3],
-        [4, 5, 6],
-      ],
-    })
-  })
-
-  test('parse parameter file with missing values', () => {
-    const data = parse(`
-[a]
-1.0
-
-[b]
-#1
-
-[c]
-2.0
-
-[d]
-#2
-
-        `)
-
-    assert.deepEqual(data, {
-      a: 1.0,
-      c: 2.0,
-    })
-  })
-
-  test('parse parameter file with strings', () => {
-    const data = parse(`
-[a]
-District_of_Columbia	Florida	Georgia
-
-[b]
-610100	United_States	Alabama
-610200	United_States	Alaska
-        `)
-
-    assert.deepEqual(data, {
-      a: ['District_of_Columbia', 'Florida', 'Georgia'],
-      b: [
-        [610100, 'United_States', 'Alabama'],
-        [610200, 'United_States', 'Alaska'],
-      ],
-    })
-  })
-
-  test('generate parameter data', () => {
-    const text = serialize({
-      'a number': 1.0,
-      'an array': [1.0, 2.0, 3.0],
-      'a matrix': [
-        [1, 2, 3],
-        [4, 5, 6],
-      ],
-    })
-
-    assert.equal(
-      text,
-      `
-[a number]
-1
-
-[an array]
-1\t2\t3
-
-[a matrix]
-1\t2\t3
-4\t5\t6
-`.trimLeft()
-    )
-  })
-
-  test('generate parameter data with missing values throws an error', () => {
-    assert.throws(() => {
-      serialize({
-        a: 1.0,
-        'some parameter': undefined,
-        c: 2.0,
-      })
-    }, "Missing value for Imperial model parameter 'some parameter'")
-  })
-
-  test('generate parameter data with strings', () => {
-    const text = serialize({
-      a: ['District_of_Columbia', 'Florida', 'Georgia'],
-      b: [
-        [610100, 'United_States', 'Alabama'],
-        [610200, 'United_States', 'Alaska'],
-      ],
-    })
-
-    assert.deepEqual(
-      text,
-      `
-[a]
-District_of_Columbia	Florida	Georgia
-
-[b]
-610100	United_States	Alabama
-610200	United_States	Alaska
-`.trimLeft()
-    )
-  })
-
   test('handles intervention strategies', () => {
     const parameters = {
       calibrationDate: '2020-03-20',
@@ -342,7 +210,7 @@ District_of_Columbia	Florida	Georgia
   })
 
   test('ensures parameters are set to 0 properly', () => {
-    const p = params.parse(`
+    const p = paramsSerialization.parse(`
   [Something something enhanced]
   9 9 9 9
   9 9 9 9
@@ -363,7 +231,7 @@ District_of_Columbia	Florida	Georgia
       calibrationDate: '2020-03-04',
     }
     params.assignParameters(p, modelParams as input.ModelParameters)
-    const output = params.serialize(p)
+    const output = paramsSerialization.serialize(p)
     assert.equal(
       output,
       `[Vary efficacies over time]
